requirements.txt
pytest
pytest-playwright
pandas
openpyxl
pytest-html-reporter
pytest-xdist